refactor : Cars, CarsTest, RacingCar, RacingCarApplication, RacingStrategy, RandomStrategy, RandomStrategyTest
Refactor 목록
- Cars, RacingCar : 일급컬렉션 내 컬렉션 불변 처리, Application 에서 racing 정책 바꿀 수 있게 설계 수정
- RacingStrategy : abstract 삭제
- RandomStrategy : java.util.Random static 처리
- RandomStrategyTest : for/if 중첩 제거
