+++ src/main/java/edu/nextstep/racingcar/car/Cars.java
@@ -1,21 +1,29 @@
 package edu.nextstep.racingcar.car;
 
-import edu.nextstep.racingcar.strategy.RandomStrategy;
+import edu.nextstep.racingcar.strategy.RacingStrategy;
 
 import java.util.ArrayList;
 import java.util.List;
 
 public class Cars {
 
-    private List<RacingCar> cars = new ArrayList<>();
+    private final List<RacingCar> cars;
 
     public Cars(List<RacingCar> racingCarList) {
         this.cars = racingCarList;
     }
 
-    public void playRacingOneRound() {
+    public static Cars createCarList(int carNum) {
+        List<RacingCar> racingCarList = new ArrayList<>();
+        for (int i = 0; i < carNum; i++) {
+            racingCarList.add(new RacingCar());
+        }
+        return new Cars(racingCarList);
+    }
+
+    public void playRacingOneRound(RacingStrategy racingStrategy) {
         for (RacingCar car : cars) {
-            car.move(new RandomStrategy());
+            car.move(racingStrategy);
         }
     }
