+++ src/main/java/edu/nextstep/racingcar/car/RacingCar.java
@@ -11,9 +11,6 @@ package edu.nextstep.racingcar.car;
 
 import edu.nextstep.racingcar.strategy.RacingStrategy;
 
-import java.util.ArrayList;
-import java.util.List;
-
 public class RacingCar {
 
     private int movedDistance = 0;
@@ -28,11 +25,5 @@ public class RacingCar {
         return movedDistance;
     }
 
-    public static Cars createCarList(int carNum) {
-        List<RacingCar> racingCarList = new ArrayList<>();
-        for (int i = 0; i < carNum; i++) {
-            racingCarList.add(new RacingCar());
-        }
-        return new Cars(racingCarList);
-    }
+
 }
