+++ src/main/java/edu/nextstep/racingcar/strategy/RacingStrategy.java
@@ -2,6 +2,6 @@ package edu.nextstep.racingcar.strategy;
 
 public interface RacingStrategy {
 
-    public abstract boolean canMove();
+    boolean canMove();
 
 }
